+++ scripts/container/preview_worker/preview_worker.sh
@@ -1,8 +1,12 @@
 #!/bin/bash
 
 # set environment variables
-MHUB_DOCKER_IMAGE_TAG="latest"
-DOCKER_BASE_IMAGE_TAG="0.2"
+MHUB_DOCKER_IMAGE_TAG=${1:-"stable"}
+MHUB_WORKER="index_worker"
+MHUB_QUEUE="index_queue"
+LOGLEVEL="INFO"
+MAPSERVER_IMAGE_TAG="0.2"
+
 GITLAB_REGISTRY_TOKEN=REDACTED_API_KEY
 AWS_ACCESS_KEY_ID="REDACTED_API_KEY"
 AWS_SECRET_ACCESS_KEY="REDACTED_API_KEY"
@@ -12,13 +16,10 @@ BROKER_IP="3.120.139.183:5672"
 MHUB_BROKER_URL=$"amqp://${BROKER_USER}:${BROKER_PW}@${BROKER_IP}//"
 MHUB_RESULT_BACKEND=$"rpc://${BROKER_USER}:${BROKER_PW}@${BROKER_IP}//"
 MP_SATELLITE_CACHE_PATH=/mnt/data/cache
-WORKER="index_worker"
-QUEUE="index_queue"
 HOST_IP=`curl http://169.254.169.254/latest/meta-data/public-ipv4`
-INDEX_OUTPUT_DIR="/mnt/data/indexes"
+MHUB_INDEX_OUTPUT_DIR="/mnt/data/indexes"
 PREVIEW_PERMALINK="http://"$HOST_IP"/geodetic.html"
 SLACK_WEBHOOK_URL="REDACTED_API_KEY"
-LOGLEVEL="INFO"
 
 # from https://gist.github.com/sj26/88e1c6584397bb7c13bd11108a579746
 function retry {
@@ -57,10 +58,10 @@ if ! [ -x "$(command -v docker)" ]; then
   sudo apt install -y htop
 fi
 
-# get rgb_worker docker container
+# get docker images
 retry 10 docker login -u gitlab-ci-token -p $GITLAB_REGISTRY_TOKEN registry.gitlab.eox.at
-retry 10 docker pull registry.gitlab.eox.at/maps/mapchete_hub/worker:$MHUB_DOCKER_IMAGE_TAG
-retry 10 docker pull registry.gitlab.eox.at/maps/docker-base/mapserver:$DOCKER_BASE_IMAGE_TAG
+retry 10 docker pull registry.gitlab.eox.at/maps/mapchete_hub/mhub:$MHUB_DOCKER_IMAGE_TAG
+retry 10 docker pull registry.gitlab.eox.at/maps/docker-base/mapserver:$MAPSERVER_IMAGE_TAG
 
 # move map directory in place
 mv map /mnt/data
@@ -76,24 +77,25 @@ docker run \
   -e AWS_ACCESS_KEY_ID=$AWS_ACCESS_KEY_ID \
   -e AWS_SECRET_ACCESS_KEY=$AWS_SECRET_ACCESS_KEY \
   -d \
-  registry.gitlab.eox.at/maps/docker-base/mapserver:$DOCKER_BASE_IMAGE_TAG
+  registry.gitlab.eox.at/maps/docker-base/mapserver:$MAPSERVER_IMAGE_TAG
 docker run \
   --rm \
-  --name $WORKER \
-  -e WORKER=$WORKER \
-  -e QUEUE=$QUEUE \
+  --name $MHUB_WORKER \
   -e AWS_ACCESS_KEY_ID=$AWS_ACCESS_KEY_ID \
   -e AWS_SECRET_ACCESS_KEY=$AWS_SECRET_ACCESS_KEY \
-  -e MHUB_BROKER_URL=$MHUB_BROKER_URL \
-  -e MHUB_RESULT_BACKEND=$MHUB_RESULT_BACKEND \
-  -e MHUB_CONFIG_DIR=$MHUB_CONFIG_DIR \
-  -e PREVIEW_PERMALINK=$PREVIEW_PERMALINK \
   -e CURL_CA_BUNDLE=/etc/ssl/certs/ca-certificates.crt \
   -e HOST_IP=$HOST_IP \
-  -e INDEX_OUTPUT_DIR=$INDEX_OUTPUT_DIR \
-  -e SLACK_WEBHOOK_URL=$SLACK_WEBHOOK_URL \
-  -e LOGLEVEL=$LOGLEVEL \
   -e LOGFILE=$LOGFILE \
+  -e LOGLEVEL=$LOGLEVEL \
+  -e MHUB_BROKER_URL=$MHUB_BROKER_URL \
+  -e MHUB_CONFIG_DIR=$MHUB_CONFIG_DIR \
+  -e MHUB_INDEX_OUTPUT_DIR=$MHUB_INDEX_OUTPUT_DIR \
+  -e MHUB_QUEUE=$MHUB_QUEUE \
+  -e MHUB_RESULT_BACKEND=$MHUB_RESULT_BACKEND \
+  -e MHUB_WORKER=$MHUB_WORKER \
+  -e PREVIEW_PERMALINK=$PREVIEW_PERMALINK \
+  -e SLACK_WEBHOOK_URL=$SLACK_WEBHOOK_URL \
   -v /mnt/data:/mnt/data \
   -d \
-  registry.gitlab.eox.at/maps/mapchete_hub/worker:$MHUB_DOCKER_IMAGE_TAG
+  registry.gitlab.eox.at/maps/mapchete_hub/mhub:$MHUB_DOCKER_IMAGE_TAG \
+  ./manage.py worker -n $MHUB_WORKER -q $MHUB_QUEUE --loglevel=$LOGLEVEL